requirements.txt
streamlit
pandas
requests
plotly
groq
python-dotenv
alpha-vantage